backend/pyproject.toml
[project]
name = "second-brain-backend"
version = "0.1.0"
description = "The Active Second Brain - AG-UI backend server with multi-agent orchestration"
requires-python = ">=3.12"
dependencies = [
    # Agent Framework + AG-UI (installs agent-framework-core, fastapi, uvicorn)
    # NOTE: requires --prerelease=allow during install (RC packages)
    "agent-framework-ag-ui",
    "agent-framework-orchestrations",
    # Azure services
    "azure-cosmos",
    "azure-identity",
    "azure-keyvault-secrets",
    # Async HTTP transport (required by azure-identity async credentials)
    "aiohttp",
    # Configuration
    "pydantic-settings",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "agent-framework-devui",
]
test = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/second_brain"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "N"]

[tool.ruff.lint.per-file-ignores]
# main.py must call load_dotenv() and configure_otel_providers() before other imports
"src/second_brain/main.py" = ["E402", "I001"]
# Document models use camelCase field names to match Cosmos DB JSON schema
"src/second_brain/models/documents.py" = ["N815"]
# Classification tools access camelCase ClassificationMeta fields
"src/second_brain/tools/classification.py" = ["N815"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
